my_gardener_project/requirements.txt
RPi.GPIO
board
adafruit-circuitpython-dht
Flask
